% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barthannwin.R
\name{barthannwin}
\alias{barthannwin}
\title{Modified Bartlett-Hann window}
\usage{
barthannwin(n)
}
\arguments{
\item{n}{Window length, specified as a positive integer.}
}
\value{
modified Bartlett-Hann window, returned as a vector.
}
\description{
Return the filter coefficients of a modified Bartlett-Hann window of length \code{n}
}
\details{
\code{barthannwin} returns an modified Bartlett-Hann window. Like Bartlett, Hann, and Hamming windows, this window has a
mainlobe at the origin and asymptotically decaying sidelobes on both sides. It is a linear combination of weighted
Bartlett and Hann windows with near sidelobes lower than both Bartlett and Hann and with far sidelobes lower than both
Bartlett and Hamming windows. The mainlobe width of the modified Bartlett-Hann window is not increased relative to
either Bartlett or Hann window mainlobes.
 
If you specify a one-point window \code{(n = 1)}, the value 1 is returned.
}
\examples{

t <- barthannwin(64)
plot (t, type = "l", xlab = "Samples", ylab =" Amplitude")

}
\author{
Original Octave code Copyright (C) 1995-2017 Andreas Weingessel \email{Andreas.Weingessel@ci.tuwien.ac.at}.
Port to R by Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
