% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttap.R
\name{buttap}
\alias{buttap}
\title{Butterworth filter prototype}
\usage{
buttap(n)
}
\arguments{
\item{n}{Order of the filter.}
}
\value{
list of class \code{'\link{Zpg}'} containg poles and gain of the filter
}
\description{
Return the poles and gain of an analog Butterworth lowpass filter prototype.
}
\details{
This function exists for Matlab/OCtave compatibility only, and is equivalent
to \code{butter(n, 1, "low", "s")}.
}
\examples{
## 9th order Butterworth low-pass analog filter
zp <- buttap(9)
w <- seq(0, 4, length.out = 128)
freqs(zp, w)

}
\author{
Original Octave code by Carnë Draug
  \email{carandraug+dev@gmail.com}. Port to R by Geert van Boxtel
  \email{G.J.M.vanBoxtel@gmail.com}.
}
