% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter.R
\name{butter}
\alias{butter}
\alias{butter.IIRfspec}
\alias{butter.default}
\title{Butterworth filter design}
\usage{
butter(n, ...)

\method{butter}{IIRfspec}(n, ...)

\method{butter}{default}(n, w, type = c("low", "high", "stop", "pass"), plane = c("z", "s"), ...)
}
\arguments{
\item{n}{filter order.}

\item{...}{additional arguments passed to butter, overriding those given by n
of class \code{IIRfspec}.}

\item{w}{critical frequencies of the filter. \code{w} must be a scalar for
low-pass and high-pass filters, and \code{w} must be a two-element vector
c(low, high) specifying the lower and upper bands in radians/second. For
digital filters, W must be between 0 and 1 where 1 is the Nyquist
frequency.}

\item{type}{filter type, one of "low", "high", "stop", or "pass".}

\item{plane}{"z" for a digital filter or "s" for an analog filter.}
}
\value{
list of class \code{'\link{Arma}'} with list elements:
\describe{
  \item{b}{moving average (MA) polynomial coefficients}
  \item{a}{autoregressive (AR) polynomial coefficients}
}
}
\description{
Compute the transfer function coefficients of a Butterworth filter
}
\details{
Butterworth filters have a magnitude response that is maximally flat in the
passband and monotonic overall. This smoothness comes at the price of
decreased rolloff steepness. Elliptic and Chebyshev filters generally provide
steeper rolloff for a given filter order.

Because butter is generic, it can be extended to accept other inputs, using
"buttord" to generate filter criteria for example.
}
\examples{
## 50 Hz notch filter
fs <- 256
bf <- butter(4, c(48, 52) / (fs / 2), "stop")
freqz(bf, fs = fs)

## EEG alpha rhythm (8 - 12 Hz) bandpass filter 
fs <- 128
fpass <- c(8, 12)
wpass <- fpass / (fs / 2)
but <- butter(5, wpass, "pass")
freqz(but, fs = fs)

## filter to remove vocals from songs, 25 dB attenuation in stop band
## (not optimal with a Butterworth filter)
fs <- 44100
specs <- buttord(230/(fs/2), 450/(fs/2), 1, 25)
bf <- butter(specs)
freqz(bf, fs = fs)
zplane(bf)

}
\references{
\url{https://en.wikipedia.org/wiki/Butterworth_filter}
}
\seealso{
\code{\link{Arma}}, \code{\link{filter}}, \code{cheby1}, \code{ellip}, \code{\link{buttord}}
}
\author{
Original Octave code by Paul Kienzle \email{pkienzle@users.sf.net},
  Doug Stewart \email{dastew@sympatico.ca}, Alexander Klein
  \email{alexander.klein@math.uni-giessen.de}, John W. Eaton. Port to R Tom
  Short, adapted by Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
