% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheb.R
\name{cheb}
\alias{cheb}
\title{Chebyshev polynomials}
\usage{
cheb(n, x)
}
\arguments{
\item{n}{Order of the polynomial, specified as a positive integer.}

\item{x}{Point or points at which to calculate the Chebyshev polynomial}
}
\value{
polynomial of order \code{x}, evaluated at point(s) \code{x}.
}
\description{
Returns the value of the \code{n}th-order Chebyshev polynomial calculated at the point \code{x}.
}
\details{
The Chebyshev polynomials are defined by the equations:
\deqn{Tn(x) = cos(n \cdot acos(x),    |x|<= 1}
\deqn{Tn(x) = cosh(n \cdot acosh(x),  |x|> 1}

If \code{x} is a vector, the output is a vector of the same size, where each element is calculated 
as \eqn{y(i) = Tn(x(i))}.
}
\examples{

cp <- cheb(5, 1)
cp <- cheb(5, c(2,3))

}
\author{
Original Octave code Copyright (C) 2002 André Carezia \email{acarezia@uol.com.br}.
Port to R by Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
