% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebwin.R
\name{chebwin}
\alias{chebwin}
\title{Chebyshev window}
\usage{
chebwin(n, at = 100)
}
\arguments{
\item{n}{Window length, specified as a positive integer.}

\item{at}{Stop-band attenuation in dB. Default: 100}
}
\value{
triangular window, returned as a vector.
}
\description{
Return the filter coefficients of a Dolph-Chebyshev window of length \code{n}
}
\details{
The window is described in frequency domain by the expression:

\deqn{W(k) = \frac{Cheb(m-1, \beta * cos(pi * k/m))}{Cheb(m-1, \beta)}}
with
\deqn{\beta = cosh(1/(m-1) * acosh(10^{(at/20)})}
and and Cheb(m,x) denoting the m-th order Chebyshev polynomial calculated at the point x.

Note that the denominator in W(k) above is not computed, and after the inverse Fourier transform the window is scaled
by making its maximum value unitary.

If you specify a one-point window \code{(n = 1)}, the value 1 is returned.
}
\examples{

cw <- chebwin(64)
plot (cw, type = "l", xlab = "Samples", ylab =" Amplitude")


}
\author{
Original Octave code Copyright (C) 2002 André Carezia \email{acarezia@uol.com.br}.
Port to R by Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
