% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmtx.R
\name{corrmtx}
\alias{corrmtx}
\title{Data matrix for autocorrelation matrix estimation}
\usage{
corrmtx(
  x,
  m,
  method = c("autocorrelation", "prewindowed", "postwindowed", "covariance", "modified")
)
}
\arguments{
\item{x}{Input data, specified as a vector.}

\item{m}{Prediction model order, specified as a positive real integer.}

\item{method}{haracter string specifying matrix computation method,
 one of:
  \describe{
   \item{\code{autocorrelation}}{Uses both prewindowed and postwindowed
   data (default). The matrix can be used to perform autoregressive 
   parameter estimation using the Yule-Walker method. For more details,
   see \code{\link{aryule}}}
   \item{\code{prewindowed}}{Uses prewindowed data only}
   \item{\code{postwindowed}}{Uses postwindowed data only}
   \item{\code{covariance}}{Uses nonwindowed data}
   \item{\code{modified}}{Uses forward and backward prediction error 
   estimates}
}}
}
\value{
A list containing the following elements:
  \describe{
   \item{H}{Data matrix, returned for autocorrelation matrix estimation. 
   The size of H depends on the matrix computation method specified in
   \code{method}}
   \item{R}{Biased autocorrelation matrix estimate \code{H’*H}}
  }
}
\description{
Given a vector \code{x} of length \code{N} and a model order \code{m},
compute the rectangular Toeplitz matrix \code{H} such that \code{H’*H} is
a biased estimate of the autocorrelation matrix. The size of \code{H}
depends on the selected method.
}
\examples{
x <- 1:5
m <- 2
HR1 <- corrmtx (x, m)
HR2 <- corrmtx (x, m, 'cov')

}
\seealso{
\code{\link{aryule}}
}
\author{
Tang Chonghao \email{chadholton@qq.com}.\cr Conversion to R by Geert
  van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
