% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplxreal.R
\name{cplxreal}
\alias{cplxreal}
\title{Sort complex conjugate pairs and real}
\usage{
cplxreal(z, tol = 100 * .Machine$double.eps, dim = 2)
}
\arguments{
\item{z}{Vector, matrix, or array of complex numbers.}

\item{tol}{Weighting factor \code{0 < tol < 1}, which determines the
tolerance of matching. Default: 100 * .Machine$double.eps.}

\item{dim}{Dimension along which to sort the complex pairs. Default: 2
(columns).}
}
\value{
A list containing two variables:
\describe{
  \item{zc}{Vector, matrix or array containg ordered complex conjugate pairs
  by increasing real parts.}
  \item{zr}{Vector, matrix or array containg ordered real numbers.}
}
}
\description{
Sort numbers into into complex-conjugate-valued and real-valued elements.
}
\details{
An error is signalled if some complex numbers could not be paired and if all
complex numbers are not exact conjugates (to within tol). Note that here is
no defined order for pairs with identical real parts but differing imaginary
parts.
}
\examples{
cplxreal(c(1, 1 + 3i, 2 - 5i, 1-3i, 2 + 5i, 4, 3))

}
\seealso{
\code{\link{cplxpair}}
}
\author{
Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
