% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtic.R
\name{filtic}
\alias{filtic}
\title{Filter Initial Conditions}
\usage{
filtic(b, a, y, x = 0)
}
\arguments{
\item{b}{the moving-average coefficients of an ARMA filter .}

\item{a}{the autoregressive (recursive) coefficients of an ARMA filter.}

\item{y}{output vector.}

\item{x}{input vector. Default: 0}
}
\value{
Initial conditions for filter with coefficients \code{a} and
  \code{b}, input vector \code{x}, and output vector \code{y}, returned as a
  vector.
}
\description{
Compute the initial conditions for a filter.
}
\details{
This function computes the same values that would be obtained from the
Matlab/Octave function \code{filter} given past inputs \code{x} and outputs
\code{y}

The vectors \code{x} and \code{y} contain the most recent inputs and outputs
respectively, with the newest values first:

\code{x = c(x(-1), x(-2), ... x(-nb)); nb = length(b)-1}\cr
\code{y = c(y(-1), y(-2), ... y(-na)); na = length(a)-a}

If \code{length(x) < nb} then it is zero padded. If \code{length(y) < na}
then it is zero padded.
}
\examples{
## Simple low pass filter
b <- c(0.25, 0.25)
a <- c(1.0, -0.5)
ic <- filtic(b, a, 1, 1)

## Simple high pass filter
b <- c(0.25, -0.25)
a <- c(1.0, 0.5)
ic <- filtic(b, a, 0, 1)

}
\seealso{
\code{\link{filter}}, \code{\link{filtfilt}}
}
\author{
David Billinghurst, \email{David.Billinghurst@riotinto.com}.\cr
Conversion to R by Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
