% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtord.R
\name{filtord}
\alias{filtord}
\alias{filtord.Arma}
\alias{filtord.Ma}
\alias{filtord.Sos}
\alias{filtord.Zpg}
\alias{filtord.default}
\title{Filter order}
\usage{
filtord(filt, ...)

\method{filtord}{Arma}(filt, ...)

\method{filtord}{Ma}(filt, ...)

\method{filtord}{Sos}(filt, ...)

\method{filtord}{Zpg}(filt, ...)

\method{filtord}{default}(filt, a = 1, ...)
}
\arguments{
\item{filt}{For the default case, the moving-average coefficients of an ARMA
filter (normally called \code{b}), specified as a numeric or complex
vector. Generically, \code{filt} specifies an arbitrary model or filter
operation.}

\item{...}{additional arguments (ignored).}

\item{a}{the autoregressive (recursive) coefficients of an ARMA filter,
specified as a numeric or complex vector. If \code{a[1]} is not equal to 1,
then filter normalizes the filter coefficients by \code{a[1]}. Therefore,
\code{a[1]} must be nonzero.}
}
\value{
Filter order, specified as a real positive scalar
}
\description{
Returns the filter order for the specified filter.
}
\details{
The filter order refers to the number of coefficients used in a digital
filter, which affects its performance and characteristics. For example, a
higher filter order typically results in a steeper roll-off and better
frequency response.

The function finds the highest power among numerator polynomial terms with
nonzero coefficients, finds the highest power for the denominator, and
returns the larger of the two. For FIR filters, the denominator equals 1, so
effectively the function returns the highest power among transfer function
polynomial terms with nonzero coefficients
}
\examples{
b <- c(1, 0)
a <- c(1, 1)
n <- filtord(b, a)

sos <- tf2sos(c(1, 0, 0, 0, 0, 0, 0, 1), c(1, 0, 0, 0, 0, 0, 0, .5))
n <- filtord(sos)

}
\seealso{
\code{\link{butter}}, \code{\link{FilterSpecs}}
}
\author{
Leonardo Araujo \email{leolca@gmail.com},\cr
 adapted by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
