% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsignal.R
\docType{package}
\name{gsignal-package}
\alias{gsignal}
\alias{gsignal-package}
\title{gsignal: Signal Processing}
\description{
R implementation of the 'Octave' package 'signal', containing a variety of signal processing tools, such as signal generation and measurement, correlation and convolution, filtering, filter design, filter analysis and conversion, power spectrum analysis, system identification, decimation and sample rate change, and windowing.
}
\details{
The package 'gsignal' is an implementation in R of the 'Octave' package
'signal'. It provides signal processing algorithms for use with R, include
the creation of waveforms, FIR and IIR filter design, spectral analysis,
Fourier and other transforms, window functions, and resampling and rate
changing.

The 'Octave' package 'signal' release 1.4.1 (2019-02-08) was used to port the
functions to R. Note that compatibility of function parameters and return
values was not explicitly pursued.
}
\references{
\url{https://wiki.octave.org/Signal_package},\cr
  \url{https://octave.sourceforge.io/signal/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gjmvanboxtel/gsignal}
  \item Report bugs at \url{https://github.com/gjmvanboxtel/gsignal/issues}
}

}
\author{
\strong{Maintainer}: Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com} (Maintainer, Author)

Authors:
\itemize{
  \item Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com} (Maintainer, Author)
  \item Tom Short \email{tshort@eprisolutions.com} (Author of 'signal' package)
  \item Octave-Forge community \email{maintainers@octave.org} (Various authors of 'Octave' functions) [contributor]
}

}
\keyword{internal}
