% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifht.R
\name{ifht}
\alias{ifht}
\title{Inverse Fast Hartley Transform}
\usage{
ifht(x, n = NROW(x))
}
\arguments{
\item{x}{input data, specified as a numeric vector or matrix. In case of a
vector it represents a single signal; in case of a matrix each column is a
signal.}

\item{n}{transform length, specified as a positive integer scalar. Default:
\code{NROW(x)}.}
}
\value{
inverse Hartley transform, returned as a vector or matrix.
}
\description{
Compute the inverse Hartley transform of a signal using the FFT
}
\details{
The Hartley transform is an integral transform closely related to the Fourier
transform, but which transforms real-valued functions to real-valued
functions. It  Compared to the Fourier transform, the Hartley transform has
the advantages of transforming real functions to real functions (as opposed
to requiring complex numbers) and of being its own inverse [1].

This function implements the inverse Hartley transform by calculating the
difference between the real- and imaginary-valued parts of the
Fourier-transformed signal [1]. The forward and inverse Hartley transforms
are the same (except for a scale factor of 1/N for the inverse Hartley
transform), but implemented using different functions.
}
\examples{
x <- 1:10
X <- fht(x)
all.equal(x, ifht(X))
# [1] TRUE

}
\references{
[1] \url{https://en.wikipedia.org/wiki/Hartley_transform}
}
\seealso{
\code{\link{fht}}, \code{\link{fft}}
}
\author{
Muthiah Annamalai, \email{muthiah.annamalai@uta.edu}; port to R by
  Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
