% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IIRfspec.R
\name{IIRfspec}
\alias{IIRfspec}
\title{IIR filter specifications}
\usage{
IIRfspec(n, Wc, type, ...)
}
\arguments{
\item{n}{filter order.}

\item{Wc}{cutoff frequency.}

\item{type}{filter type, normally one of "low", "high", "stop", or "pass".}

\item{...}{other filter description characteristics, possibly including Rp
for dB of pass band ripple or Rs for dB of stop band ripple, depending on
filter type (Butterworth, Chebyshev, etc.).}
}
\value{
A list of class \code{'IIRfspec'} with the following list elements
  (repeats of the input arguments:
\describe{
  \item{n}{filter order}
  \item{Wc}{cutoff frequency}
  \item{type}{filter type, normally one of "low", "high", "stop", or "pass".}
  \item{...}{other filter description characteristics, possibly including Rp
  for dB of pass band ripple or Rs for dB of stop band ripple, depending on
  filter type (Butterworth, Chebyshev, etc.).}
}
}
\description{
IIR filter specifications, including order, frequency cutoff, type, and
possibly others.
}
\examples{
filt <- IIRfspec(3, 0.1, "low")

}
\seealso{
\code{\link{filter}}, butter and buttord, cheby1 and cheb1ord, ellip and ellipord.
}
\author{
Tom Short \email{tshort@eprisolutions.com}, adapted by Geert van
  Boxtel \email{gjmvanboxtel@gmail.com}
}
