% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isallpass.R
\name{isallpass}
\alias{isallpass}
\alias{isallpass.Arma}
\alias{isallpass.Ma}
\alias{isallpass.Sos}
\alias{isallpass.Zpg}
\alias{isallpass.default}
\title{Determine whether a digital filter is allpass}
\usage{
isallpass(filt, ...)

\method{isallpass}{Arma}(filt, ...)

\method{isallpass}{Ma}(filt, ...)

\method{isallpass}{Sos}(filt, ...)

\method{isallpass}{Zpg}(filt, ...)

\method{isallpass}{default}(filt, a = 1, ...)
}
\arguments{
\item{filt}{For the default case, the moving-average coefficients of an ARMA
filter (normally called \code{b}), specified as a numeric or complex
vector. Generically, \code{filt} specifies an arbitrary model or filter
operation.}

\item{...}{additional arguments (ignored).}

\item{a}{the autoregressive (recursive) coefficients of an ARMA filter,
specified as a numeric or complex vector. If \code{a[1]} is not equal to 1,
then filter normalizes the filter coefficients by \code{a[1]}. Therefore,
\code{a[1]} must be nonzero.}
}
\value{
Logical indicating whether the filter is allpass
}
\description{
An all-pass filter is a filter that passes all frequencies equally in gain,
but changes the phase relationship among various frequencies.
}
\examples{
a <- c(1, 2, 3)
b <- c(3, 2, 1)
isallpass (b, a)

# H(z) = (b1 - z^-1) * (b2 - z^-1) / ((1 - b1*z^-1) * (1 - b2*z^-1))
b1 <- 0.5 * (1 + 1i)
b2 <- 0.7 * (cos(pi / 6) + 1i*sin(pi / 6))
b <- conv(c(b1, -1), c(b2, -1))
a <- conv(c(1, (-1) * Conj(b1)), c(1, (-1) * Conj(b2)))
freqz (b, a)
isallpass (b, a)
 
}
\references{
[1] Shyu, Jong-Jy, & Pei, Soo-Chang,
A new approach to the design of complex all-pass IIR digital filters,
Signal Processing, 40(2–3), 207–215, 1994.
https://doi.org/10.1016/0165-1684(94)90068-x

[2] Vaidyanathan, P. P. Multirate Systems and Filter Banks.
1st edition, Pearson College Div, 1992.
}
\author{
Leonardo Araujo \email{leolca@gmail.com},\cr
 conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
