% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismaxphase.R
\name{ismaxphase}
\alias{ismaxphase}
\alias{ismaxphase.Arma}
\alias{ismaxphase.Ma}
\alias{ismaxphase.Sos}
\alias{ismaxphase.Zpg}
\alias{ismaxphase.default}
\title{Determine whether a digital filter is maximum phase}
\usage{
ismaxphase(filt, ...)

\method{ismaxphase}{Arma}(filt, ...)

\method{ismaxphase}{Ma}(filt, ...)

\method{ismaxphase}{Sos}(filt, ...)

\method{ismaxphase}{Zpg}(filt, ...)

\method{ismaxphase}{default}(filt, a = 1, tol = .Machine$double.eps^(3/4), ...)
}
\arguments{
\item{filt}{For the default case, the moving-average coefficients of an ARMA
filter (normally called \code{b}), specified as a numeric or complex
vector. Generically, \code{filt} specifies an arbitrary model or filter
operation.}

\item{...}{additional arguments (ignored).}

\item{a}{the autoregressive (recursive) coefficients of an ARMA filter,
specified as a numeric or complex vector. If \code{a[1]} is not equal to 1,
then filter normalizes the filter coefficients by \code{a[1]}. Therefore,
\code{a[1]} must be nonzero.}

\item{tol}{tolerance to determine when two numbers are close enough to be
considered equal. Default: .Machine$double.eps^(3 / 4)}
}
\value{
Logical indicating whether the filter is maximum phase
}
\description{
A maximum-phase filter is a type of causal filter that has all its zeros
outside the unit circle in the z-plane, so its numerator coefficients b are
typically larger in magnitude than the denominator coefficients a, resulting
in the largest group delay among all causal filters with the same magnitude
response.
}
\examples{

# maximum-phase FIR filter
fir <- Ma(c(0.5, -1))
zplane(fir)
ismaxphase(fir)

#  maximum-phase IIR filter
iir <- Arma(conv(c(1, -1.2), c(1, -1.5)), conv(c(1, -0.6), c(1, -0.7))) 
zplane(iir)
ismaxphase(iir)

sos <- Sos(c(1, -2.7, 1.8, 1, -1.3, 0.42))
zplane(sos)
ismaxphase(sos)
 
}
\references{
Oppenheim, A., and Schafer, R. Discrete-Time Signal Processing.
3rd edition, Pearson, 2009
}
\seealso{
\code{\link{isminphase}}
}
\author{
Leonardo Araujo \email{leolca@gmail.com},\cr
 conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
