% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isminphase.R
\name{isminphase}
\alias{isminphase}
\alias{isminphase.Arma}
\alias{isminphase.Ma}
\alias{isminphase.Sos}
\alias{isminphase.Zpg}
\alias{isminphase.default}
\title{Determine whether a digital filter is minimum phase}
\usage{
isminphase(filt, ...)

\method{isminphase}{Arma}(filt, ...)

\method{isminphase}{Ma}(filt, ...)

\method{isminphase}{Sos}(filt, ...)

\method{isminphase}{Zpg}(filt, ...)

\method{isminphase}{default}(filt, a = 1, tol = .Machine$longdouble.eps^(3/4), ...)
}
\arguments{
\item{filt}{For the default case, the moving-average coefficients of an ARMA
filter (normally called \code{b}), specified as a numeric or complex
vector. Generically, \code{filt} specifies an arbitrary model or filter
operation.}

\item{...}{additional arguments (ignored).}

\item{a}{the autoregressive (recursive) coefficients of an ARMA filter,
specified as a numeric or complex vector. If \code{a[1]} is not equal to 1,
then filter normalizes the filter coefficients by \code{a[1]}. Therefore,
\code{a[1]} must be nonzero.}

\item{tol}{tolerance to determine when two numbers are close enough to be
considered equal. Default: .Machine$longdouble.eps^(3 / 4)}
}
\value{
Logical indicating whether the filter is minimum phase
}
\description{
A minimum-phase filter is a type of causal filter that has all its poles and
zeros inside the unit circle in the z-plane. Note that minimum-phase filters
are stable by definition since the poles must be inside the unit circle. In
addition, because the zeros must also be inside the unit circle, the inverse
filter is also stable when the filter is minimum phase.
}
\examples{

b <- c(3, 1)
a <- c(1, .5)
zplane(b, a)
isminphase(b, a)

zp <- butter(6, 0.25, output="Zpg")
zplane(zp)
isminphase(zp)
  
}
\references{
Oppenheim, A., and Schafer, R. Discrete-Time Signal Processing.
3rd edition, Pearson, 2009
}
\seealso{
\code{\link{ismaxphase}}
}
\author{
Leonardo Araujo \email{leolca@gmail.com},\cr
 conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
