% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isstable.R
\name{isstable}
\alias{isstable}
\alias{isstable.Arma}
\alias{isstable.Ma}
\alias{isstable.Sos}
\alias{isstable.Zpg}
\alias{isstable.default}
\title{Determine whether a digital filter is stable}
\usage{
isstable(filt, ...)

\method{isstable}{Arma}(filt, ...)

\method{isstable}{Ma}(filt, ...)

\method{isstable}{Sos}(filt, ...)

\method{isstable}{Zpg}(filt, ...)

\method{isstable}{default}(filt, a = 1, ...)
}
\arguments{
\item{filt}{For the default case, the moving-average coefficients of an ARMA
filter (normally called \code{b}), specified as a numeric or complex
vector. Generically, \code{filt} specifies an arbitrary model or filter
operation.}

\item{...}{additional arguments (ignored).}

\item{a}{the autoregressive (recursive) coefficients of an ARMA filter,
specified as a numeric or complex vector. If \code{a[1]} is not equal to 1,
then filter normalizes the filter coefficients by \code{a[1]}. Therefore,
\code{a[1]} must be nonzero.}
}
\value{
Logical indicating whether the filter is stable
}
\description{
A stable digital filter is one where the impulse response approaches zero 
as time goes to infinity, ensuring that the output remains bounded for any
bounded input. Stability is typically guaranteed if all poles of the
filter's transfer function are located inside the unit circle in the
complex plane.
}
\examples{

b <- c(1, 2, 3, 4, 5, 5, 1, 2)
a <- c(4, 5, 6, 7, 9, 10, 4, 6)
zplane(b, a)
isstable (b, a)

zp <- butter(6, 0.7, 'high', output = 'Zpg')
zplane(zp)
isstable (zp)
 
}
\author{
Vasilis Lefkopoulos \email{vlefkopo@gmail.com},\cr
 conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
