% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaiser.R
\name{kaiser}
\alias{kaiser}
\title{Kaiser window}
\usage{
kaiser(n, beta = 0.5)
}
\arguments{
\item{n}{Window length, specified as a positive integer.}

\item{beta}{Shape factor, specified as a positive real scalar. The parameter \code{beta} affects the sidelobe attenuation
of the Fourier transform of the window. Default: 0.5}
}
\value{
kaiser window, returned as a vector.
}
\description{
Return the filter coefficients of a kaiser window of length \code{n}
}
\details{
The Kaiser, or Kaiser-Bessel, window is a simple approximation of the DPSS window using Bessel functions, discovered
by James Kaiser.
\deqn{w(x) = frac{besselI(0, \beta * \sqrt{(1 - (2*x/m)^{2}))}}{besseli(0, \beta)} , m/2 <= x <= m/2}
Variable parameter \eqn{\beta} determines the tradeoff between main lobe width and side lobe levels of the spectral
leakage pattern. Increasing \eqn{\beta} widens the mainlobe and decreases the amplitude of the sidelobes
(i.e., increases the attenuation).
}
\examples{

k <- kaiser(200, 2.5)
plot (k, type = "l", xlab = "Samples", ylab =" Amplitude")

}
\author{
Original Octave code Copyright (C) 1995, 1996, 1997 Kurt Hornik \email{Kurt.Hornik@ci.tuwien.ac.at},
Copyright (C) 2000 Paul Kienzle \email{pkienzle@users.sf.net}. Port to R by Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
