% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcumq.R
\name{marcumq}
\alias{marcumq}
\title{Marcum Q function}
\usage{
marcumq(a, b, m = 1)
}
\arguments{
\item{a, b}{input arguments, specified as non-negative real numbers.}

\item{m}{order, specified as a positive integer}
}
\value{
Marcum Q function.
}
\description{
Compute the generalized Marcum Q function
}
\details{
The code for this function was taken from the help file of the \code{cdfkmu}
function in the \code{lmomco} package, based on a suggestion of Daniel
Wollschlaeger.
}
\examples{
mq <- marcumq(12.4, 12.5)

}
\references{
\url{https://cran.r-project.org/package=lmomco}
}
\author{
William Asquith, \email{william.asquith@ttu.edu}.
}
