% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasez.R
\name{phasez}
\alias{phasez}
\alias{phasez.default}
\alias{phasez.Arma}
\alias{phasez.Ma}
\alias{phasez.Sos}
\alias{phasez.Zpg}
\alias{print.phasez}
\alias{summary.phasez}
\alias{print.summary.phasez}
\alias{phasez_plot}
\title{Phase response of digital filter}
\usage{
phasez(filt, ...)

\method{phasez}{default}(
  filt,
  a = 1,
  n = 512,
  whole = ifelse((is.numeric(filt) && is.numeric(a)), FALSE, TRUE),
  fs = 2 * pi,
  ...
)

\method{phasez}{Arma}(filt, ...)

\method{phasez}{Ma}(filt, ...)

\method{phasez}{Sos}(filt, ...)

\method{phasez}{Zpg}(filt, n = 512, whole = FALSE, fs = 2 * pi, ...)

\method{print}{phasez}(x, ...)

\method{summary}{phasez}(object, ...)

\method{print}{summary.phasez}(x, ...)

phasez_plot(w, phi, ...)
}
\arguments{
\item{filt}{for the default case, the moving-average coefficients of an ARMA
model or filter. Generically, \code{filt} specifies an arbitrary model or
filter operation.}

\item{...}{for methods of \code{phasez}, arguments are passed to the default
method. For \code{phasez_plot}, additional arguments are passed through to
plot.}

\item{a}{the autoregressive (recursive) coefficients of an ARMA filter.}

\item{n}{number of points at which to evaluate the frequency response. If
\code{n} is a vector with a length greater than 1, then evaluate the
frequency response at these points. For fastest computation, \code{n}
should factor into a small number of small primes. Default: 512.}

\item{whole}{FALSE (the default) to evaluate around the upper half of the
unit circle or TRUE to evaluate around the entire unit circle.}

\item{fs}{sampling frequency in Hz. If not specified (default = 2 * pi), the
frequencies are in radians.}

\item{x}{object to be printed or plotted.}

\item{object}{object of class \code{"phasez"} for \code{summary}}

\item{w}{vector of frequencies}

\item{phi}{phase response, specified as a vector.}
}
\value{
For \code{phasez}, a list of class \code{'phasez'} with items:
\describe{
  \item{phi}{complex array of phase responses at frequencies \code{w}.}
  \item{w}{array of frequencies.}
  \item{u}{units of (angular) frequency; either rad/s or Hz.}
}
}
\description{
Compute the z-plane frequency response of an ARMA model or rational IIR
filter.
}
\examples{
b <- c(1, 0, -1)
a <- c(1, 0, 0, 0, 0.25)
phasez(b, a)

ph <- phasez(b, a)
summary(ph)

sos <- cheby2(40, 50, 0.4, output = "Sos")
ph <- phasez(sos)
summary(ph)
ph

}
\references{
Oppenheim, A., and Schafer, R. Discrete-Time Signal
  Processing. 3rd edition, Pearson, 2009.
}
\author{
Leonardo Araujo \email{leolca@gmail.com}.\cr
 Port to R by Geert van Boxtel, \email{gjmvanboxtel@gmail.com}
}
