% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Change sampling rate}
\usage{
resample(x, p, q, h)
}
\arguments{
\item{x}{input data, specified as a numeric vector or matrix. In case of a
vector it represents a single signal; in case of a matrix each column is a
signal.}

\item{p, q}{resampling factors, specified as positive integers. \code{p / q}
is the resampling factor.}

\item{h}{Impulse response of the FIR filter specified as a numeric vector or
matrix. If it is a vector, then it represents one FIR filter to may be
applied to multiple signals in \code{x}; if it is a matrix, then each
column is a separate FIR impulse response.}
}
\value{
output signal, returned as a vector or matrix. Each column has length
  \code{ceiling(((length(x) - 1) * p + length(h)) / q)}..
}
\description{
Resample using a polyphase algorithm.
}
\examples{
tx <- seq(0, 360 - 3, 6)
x <- sin(2 * pi * tx / 120)
ty <- seq(0, 360 - 2, 4)
y <- resample(x, 3, 2)
tz <- seq(0, 360 - 4, 9)
z <- resample(x, 2, 3)
plot(tx, x, type = "b", pch = 1)
points(ty, y, col="red")
points(tz, z, col = "blue")

}
\references{
[1] Proakis, J.G., and Manolakis, D.G. (2007).
Digital Signal Processing: Principles, Algorithms, and Applications,
4th ed., Prentice Hall, Chap. 6.\cr
[2] Oppenheim, A.V., Schafer, R.W., and Buck, J.R. (1999).
Discrete-time signal processing, Signal processing series,
Prentice-Hall.
}
\author{
Original Octave code by Eric Chassande-Mottin
  \email{ecm@apc.univ-paris7.fr}, port to R by Geert van Boxtel
  \email{G.J.M.vanBoxtel@gmail.com}.
}
