% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{roots}
\alias{roots}
\title{Polynomial roots}
\usage{
roots(p, method = c("polyroot", "eigen"))
}
\arguments{
\item{p}{Polynomial coefficients with coefficients given in order from
highest to lowest polynomial power. This is the Matlab/Octave convention;
it is opposite of the convention used by polyroot.}

\item{method}{Either “polyroot” (default) which uses \code{polyroot} for its
computations internally (and should be more accurate) or “eigen” which
uses eigenvalues of the companion matrix for its computation. The latter
returns complex values in case of real valued solutions in less cases.}
}
\value{
A complex vector with the roots of the polynomial (converted to
  numeric if all imaginary components are 0).
}
\description{
Compute the roots of a polynomial represented by \code{p}
}
\details{
The \code{roots} function solves polynomial equations of the form \code{p(1)
* x^(n-1) + ... + p(n-1) * x + p(n) = 0}. Polynomial equations contain a
single variable with nonnegative exponents.
}
\examples{
roots(1:3)
polyroot(3:1) # should be the same
poly(roots(1:3))

roots(1:3, method="eigen") # using eigenvalues

}
\seealso{
\code{\link{poly}}
}
\author{
Original Octave version by Kurt Hornik. Conversion to R by Tom Short.
  Adapted by Geert van Boxtel
}
