% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgolayfilt.R
\name{filter.sgolayFilter}
\alias{filter.sgolayFilter}
\alias{sgolayfilt}
\title{Savitzky-Golay filtering}
\usage{
\method{filter}{sgolayFilter}(filt, x, ...)

sgolayfilt(x, p = 3, n = p + 3 - p\%\%2, m = 0, ts = 1)
}
\arguments{
\item{filt}{Filter characteristics, usually the result of a call to \code{sgolay}}

\item{x}{Polynomial filter order; must be smaller than \code{n}.}

\item{...}{Additional arguments (ignored)}

\item{p}{Polynomial filter order; must be smaller than \code{n}.}

\item{n}{Filter length; must a an odd positive integer.}

\item{m}{Return the m-th derivative of the filter coefficients. Default: 0}

\item{ts}{Scaling factor. Default: 1}
}
\value{
The filtered signal, returned as a vector.
}
\description{
Filter a signal with a Savitzky-Golay FIR filter
}
\details{
Savitzky-Golay smoothing filters are typically used to "smooth out" a noisy
signal whose frequency span (without noise) is large. They are also called
digital smoothing polynomial filters or least-squares smoothing filters.
Savitzky-Golay filters perform better in some applications than standard
averaging FIR filters, which tend to filter high-frequency content along with
the noise. Savitzky-Golay filters are more effective at preserving high
frequency signal components but less successful at rejecting noise.

Savitzky-Golay filters are optimal in the sense that they minimize the
least-squares error in fitting a polynomial to frames of noisy data.
}
\examples{
# Compare a 5 sample averager, an order-5 butterworth lowpass
# filter (cutoff 1/3) and sgolayfilt(x, 3, 5), the best cubic
# estimated from 5 points.
#bf <- butter(5, 1/3)
b <- c(0.01061191, 0.05305957, 0.10611913, 0.10611913, 0.05305957, 0.01061191)
a <- c(1.00000000, -1.64484891,  1.58661518, -0.80488189,  0.22994912, -0.02725227)
bf <- Arma(b, a)
x <- c(rep(0, 15), rep(10, 10), rep(0, 15))
sg <- sgolayfilt(x)
plot(sg, type="l", xlab = "", ylab = "")
lines(filtfilt(rep(1, 5) / 5, 1, x), col = "red") # averaging filter
lines(filtfilt(bf, x), col = "blue")              # butterworth
points(x, pch = "x")                              # original data
legend("topleft", c("sgolay (3,5)", "5 sample average", "order 5
Butterworth", "original data"), lty=c(1, 1, 1, NA),
pch = c(NA, NA, NA, "x"), col = c(1, "red", "blue", 1))

}
\seealso{
\code{\link{sgolay}}
}
\author{
Paul Kienzle \email{pkienzle@users.sf.net}, port to R Tom Short.
  Minor changes by Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
