% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sosfilt.R
\name{sosfilt}
\alias{sosfilt}
\title{Second-order filtering}
\usage{
sosfilt(sos, x)
}
\arguments{
\item{sos}{Second-order section representation, specified as an nrow-by-6
matrix, whose rows contain the numerator and denominator coefficients of
the second-order sections:\cr \code{sos <- rbind(cbind(B1, A1), cbind(...),
cbind(Bn, An))}, where \code{B1 <- c(b0, b1, b2)}, and \code{A1 <- c(a0,
a1, a2)} for section 1, etc. The b0 entry must be nonzero for each section.}

\item{x}{The input data to be filtered, coerced to a vector.}
}
\value{
The filtered signal, normally of the same length as the input signal
  \code{x}, returned as a vector
}
\description{
One-dimensional second-order (biquadratic) IIR digital filtering
}
\examples{
fs <- 1000                                           # sampling frequency
t <- seq(0, 1, 1/fs)                                 # 1 second sample
s <- sin(2* pi * t * 6)                              # 6 Hz sinus
x <- s + rnorm(length(t))                            # add noise
plot(t, x, type = "l", col="light gray")
lines(t, s, col="black")
#bf <- signal::butter(3, 0.02)                        # low-pass 0.02 * 1000 = 20 Hz
bf <- Arma(c(2.914649e-05, 8.743948e-05, 8.743948e-05, 2.914649e-05),
           c(1.0000000, -2.8743569,  2.7564832, -0.8818931))
sosg <- as.Sos(bf)                                   # convert to second order sections
sos <- sosg$sos
sos[1, 1:3] <- sos[1, 1:3] * sosg$g                  # apply gain factor
y <- sosfilt(matrix(sos, ncol=6), x)                 # apply filter
lines(t, y, col="red")
yy <- filtfilt(sosg, x)                              # filtfilt and filter handle gain factor
lines(t, yy, col="blue")

}
\seealso{
\code{\link{filter}}, \code{\link{filtfilt}}
}
\author{
Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com}.
}
