% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.R
\name{unwrap}
\alias{unwrap}
\title{Unwrap phase angles}
\usage{
unwrap(x, tol = pi, MARGIN = 2)
}
\arguments{
\item{x}{Input array, specified as a vector, matrix, or multidimensional array.}

\item{tol}{Jump threshold to apply phase shift, specified as a scalar. A jump
threshold less than \eqn{pi} has the same effect as the threshold
\eqn{pi}. Default: \deqn{pi}.}

\item{MARGIN}{A vector giving the subscripts which the function will be
applied over. E.g., for a matrix 1 indicates rows, 2 indicates columns,
c(1, 2) indicates rows and columns. Where x has named dimnames, it can be a
character vector selecting dimension names. If MARGIN is larger than the
dimensions of x, the result will have MARGIN dimensions. Default: 2
(columns).}
}
\value{
Unwraped phase angle, returned as a vector, matrix, or
  multidimensional array. The size of the output is always the same as the
  size of the input.
}
\description{
Unwrap radian phases by adding or subtracting multiples of \code{2 * pi}.
}
\examples{
## Define spiral shape.
t <- seq(0, 6 * pi, length.out = 201)
x <- t / pi * cos(t)
y <- t / pi * sin(t)
plot(x, y, type = "l")
## find phase angle
p = atan2(y, x)
plot(t, p, type="l")
## unwrap it
q = unwrap(p)
plot(t, q, type ="l")

}
\author{
Bill Lash, port to R by Geert van Boxtel
  \email{gjmvanboxtel@gmail.com}
}
