% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vco.R
\name{vco}
\alias{vco}
\title{Voltage-controlled oscillator}
\usage{
vco(x, fc = fs/4, fs = 1)
}
\arguments{
\item{x}{Input data, specified as a real vector or real matrix. \code{x}
ranges from –1 to 1, where \code{x = –1} corresponds to 0 frequency output,
\code{x = 0} corresponds to \code{fc}, and \code{x = 1} corresponds to
\code{2 * fc}. If \code{x} is a matrix, \code{vco} produces a matrix whose
columns oscillate according to the columns of \code{x}.}

\item{fc}{Carrier frequency used to modulate the input signal, specified as a
real positive scalar, or a vector of 2 positive scalars. In the latter
case, \code{fc} give the frequency modulation range limits. Values must be
in the range \code{0} to \code{fs / 2}. Default: \code{fs / 4}.}

\item{fs}{Sampling rate, specified as a positive scalar. Default: 1}
}
\value{
Output signal, returned as a real vector or real matrix. It is the
  same size as \code{x} and has amplitude equal to 1.
}
\description{
Creates a signal that oscillates at a frequency determined by input \code{x}
with a sampling frequency fs.
}
\examples{

x <- c(0, 0.5, 0.8, -0.2)
y1 <- vco (x, 10, 100)
y2 <- vco (x, c(5, 15), 100)

fs <- 10000
t <- seq(0, 2, 1 / fs)
x <- sawtooth(2 * pi * t, 0.75)
y <- vco(x, c(0.1, 0.4) * fs, fs)
specgram(y, fs = fs)

}
\seealso{
\code{\link{pulstran}}
}
\author{
The Octave Project Developers, \email{maintainers@octave.org}, 
  Conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
