% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zpg.R
\name{Zpg}
\alias{Zpg}
\alias{as.Zpg}
\alias{as.Zpg.Arma}
\alias{as.Zpg.Ma}
\alias{as.Zpg.Zpg}
\title{Zero pole gain model}
\usage{
Zpg(zero, pole, gain)

as.Zpg(x, ...)

## S3 method for class 'Arma'
as.Zpg(x, ...)

## S3 method for class 'Ma'
as.Zpg(x, ...)

## S3 method for class 'Zpg'
as.Zpg(x, ...)
}
\arguments{
\item{zero}{complex vector of the zeros of the model.}

\item{pole}{complex vector of the poles of the model.}

\item{gain}{gain of the model.}

\item{x}{model to be converted.}

\item{...}{additional arguments (ignored).}
}
\value{
A list of class Zpg with the following list elements:
\describe{
  \item{zero}{complex vector of the zeros of the model}
  \item{pole}{complex vector of the poles of the model}
  \item{gain}{gain of the model}
}
}
\description{
Create an zero pole gain model of an ARMA filter
}
\details{
\code{as.Zpg} converts from other forms, including \code{Arma} and \code{Ma}.
}
\examples{
filt <- Zpg(c(-1, -1), -1, 1/3)
#zplane(filt)

}
\seealso{
See also \code{\link{Arma}}
}
\author{
Tom Short \email{tshort@eprisolutions.com}
}
