% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsignal.R
\docType{package}
\name{gsignal-package}
\alias{gsignal}
\alias{gsignal-package}
\title{gsignal: Signal Processing}
\description{
Gsignal is my own R implementation of the Matlab/Octave signal package.
}
\details{
gsignal is my own implementation of the Matlab/Octave package signal
}
\references{
To be filled in
}
\keyword{internal}
