% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsignal.R
\docType{package}
\name{gsignal-package}
\alias{gsignal}
\alias{gsignal-package}
\title{gsignal: Signal Processing}
\description{
I wanted to learn more about signal analysis, so I thought it would be useful to study
    the Matlab/Octave signal package. Gsignal is my own R implementation of the Octave signal package.
    This allowed me to get a basic understanding of the algorithms and experiment with them in R.
}
\details{
gsignal is my own implementation of the Matlab/Octave package signal
}
\references{
To be filled in
}
\keyword{internal}
