% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsignal.R
\docType{package}
\name{gsignal-package}
\alias{gsignal}
\alias{gsignal-package}
\title{gsignal: Signal Processing}
\description{
R implementation of the 'Octave' package 'signal', containing a variety of signal processing tools, such as signal generation and measurement, correlation and convolution, filtering, filter design, filter analysis and conversion, power spectrum analysis, system identification, decimation and sample rate change, and windowing.
}
\details{
The package 'gsignal' is an implementation in R of the 'Octave' package
'signal'. It provides signal processing algorithms for use with R, include
the creation of waveforms, FIR and IIR filter design, spectral analysis,
Fourier and other transforms, window functions, and resampling and rate
changing.

The 'Octave' package 'signal' release 1.4.1 (2019-02-08) was used to port the
functions to R. Note that compatibility of function parameters and return
values was not explicitly pursued.
}
\references{
\url{https://wiki.octave.org/Signal_package},\cr
  \url{https://octave.sourceforge.io/signal/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gjmvanboxtel/gsignal}
  \item Report bugs at \url{https://github.com/gjmvanboxtel/gsignal/issues}
}

}
\author{
\strong{Maintainer}: Geert van Boxtel \email{G.J.M.vanBoxtel@gmail.com} (Maintainer)

Authors:
\itemize{
  \item Tom Short \email{tshort@eprisolutions.com} (Author of 'signal' package)
  \item Paul Kienzle (Majority of the original sources)
}

Other contributors:
\itemize{
  \item Ben Abbott [contributor]
  \item Juan Aguado [contributor]
  \item Muthiah Annamalai [contributor]
  \item Leonardo Araujo [contributor]
  \item William Asquith [contributor]
  \item David Bateman [contributor]
  \item David Billinghurst [contributor]
  \item Juan Pablo Carbajal [contributor]
  \item André Carezia [contributor]
  \item Vincent Cautaerts [contributor]
  \item Eric Chassande-Mottin [contributor]
  \item Luca Citi [contributor]
  \item Dave Cogdell [contributor]
  \item Carlo de Falco [contributor]
  \item Carne Draug [contributor]
  \item Pascal Dupuis [contributor]
  \item John W. Eaton [contributor]
  \item R.G.H Eschauzier [contributor]
  \item Andrew Fitting [contributor]
  \item Alan J. Greenberger [contributor]
  \item Mike Gross [contributor]
  \item Daniel Gunyan [contributor]
  \item Kai Habel [contributor]
  \item Kurt Hornik [contributor]
  \item Jake Janovetz [contributor]
  \item Alexander Klein [contributor]
  \item Peter V. Lanspeary [contributor]
  \item Bill Lash [contributor]
  \item Friedrich Leissh [contributor]
  \item Laurent S. Mazet [contributor]
  \item Mike Miller [contributor]
  \item Petr Mikulik [contributor]
  \item Paolo Neis [contributor]
  \item Georgios Ouzounis [contributor]
  \item Sylvain Pelissier [contributor]
  \item Francesco Potortì [contributor]
  \item Charles Praplan [contributor]
  \item Lukas F. Reichlin [contributor]
  \item Tony Richardson [contributor]
  \item Asbjorn Sabo [contributor]
  \item Thomas Sailer [contributor]
  \item Rolf Schirmacher [contributor]
  \item Rolf Schirmacher [contributor]
  \item Ivan Selesnick [contributor]
  \item Julius O. Smith III [contributor]
  \item Peter L. Soendergaard [contributor]
  \item Quentin Spencer [contributor]
  \item Doug Stewart [contributor]
  \item P. Sudeepam [contributor]
  \item Stefan van der Walt [contributor]
  \item Andreas Weber [contributor]
  \item P. Sudeepam [contributor]
  \item Andreas Weingessel [contributor]
}

}
\keyword{internal}
