% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impzlength.R
\name{impzlength}
\alias{impzlength}
\alias{impzlength.Arma}
\alias{impzlength.Ma}
\alias{impzlength.Sos}
\alias{impzlength.Zpg}
\alias{impzlength.default}
\title{Impulse response length}
\usage{
impzlength(filt, ...)

\method{impzlength}{Arma}(filt, ...)

\method{impzlength}{Ma}(filt, ...)

\method{impzlength}{Sos}(filt, ...)

\method{impzlength}{Zpg}(filt, ...)

\method{impzlength}{default}(filt, a = 1, tol = 5e-05, ...)
}
\arguments{
\item{filt}{For the default case, the moving-average coefficients of an ARMA
filter (normally called \code{b}), specified as a numeric or complex
vector. Generically, \code{filt} specifies an arbitrary model or filter
operation.}

\item{...}{additional arguments (ignored).}

\item{a}{the autoregressive (recursive) coefficients of an ARMA filter,
specified as a numeric or complex vector. If \code{a[1]} is not equal to 1,
then filter normalizes the filter coefficients by \code{a[1]}. Therefore,
\code{a[1]} must be nonzero.}

\item{tol}{specifies the tolerance used to estimate the effective length of
an IIR filter's impulse response. The default tolerance is 5e-5.
Increasing @var{tol} estimates a shorter effective length, while decreasing
@var{tol} produces a longer effective length.}
}
\value{
Length of the impulse response, specified as a positive integer. For
  stable IIR filters with absolutely summable impulse responses, impzlength
  returns an effective length for the impulse response beyond which the
  coefficients are essentially zero. You can control this cutoff point by
  specifying the optional tol input argument.
}
\description{
For a finite impulse response (FIR) filter specified by the numerator
 coefficients \code{b}, the length is simply the number of coefficients in
 \code{b}.
}
\details{
For an infinite impulse response (IIR) filter specified by the numerator
\code{b} and denominator \code{a} polynomials in z^-1, the function computes
an effective impulse response sequence length.


The algorithm proceeds as follows:
\itemize{
   \item{If the filter is FIR, the length is simply length (b)}.
   \item{The poles of the transfer function are computed as the roots
    of the denominator polynomial \code{a}.}
   \item{The multiplicity of the dominant pole (the pole with the largest
   magnitude) is determined by counting poles at the same complex coordinate
   within tolerance.}
   \item{For a stable IIR filter (dominant pole magnitude < 1 - 10^{-5}),
    the effective length is estimated as
    
    \code{floor (M * log10 (tol) / log10 (maxpole)) + delay}
    
    where M is the multiplicity of the dominant pole and d is the initial
     delay (number of leading zeros in b).}
   \item{For an unstable IIR filter (dominant pole magnitude > 1 + 10^{-4}),
    a heuristic formula is used:
    
    \code{floor (6 / log10 (maxpole))}}
   \item{For filters with poles near the unit circle (oscillatory behavior),
    the length is the maximum of: five periods of the slowest oscillation,
     and the decay length of damped poles, plus the initial delay. }
}
}
\examples{
b <- 1
a <- c(1, -0.9)
len <- impzlength(b, a)

}
\author{
Tang Chonghao, \email{chadholton@qq.com}.\cr Conversion to R by
  Geert van Boxtel, \email{gjmvanboxtel@gmail.com}
}
