% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyreduce.R
\name{polyreduce}
\alias{polyreduce}
\title{Reduce polynomial}
\usage{
polyreduce(pc)
}
\arguments{
\item{pc}{vector of polynomial coefficients}
}
\value{
Vector of reduced polynomial coefficients.
}
\description{
Reduce a polynomial coefficient vector to a minimum number of terms by
stripping off any leading zeros.
}
\examples{
p <- polyreduce(c(0, 0, 1, 2, 3))

}
\author{
Tony Richardson, \email{arichard@stark.cc.oh.us},\cr
 adapted by John W. Eaton.\cr
  Conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}
}
