% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow2db.R
\name{pow2db}
\alias{pow2db}
\alias{db2pow}
\title{Power - decibel conversion}
\usage{
pow2db(x)

db2pow(x)
}
\arguments{
\item{x}{input data, specified as a numeric vector, matrix, or
multidimensional array. Must be non-negative for numeric \code{x}.}
}
\value{
Converted data, same type and dimensions as \code{x}.
}
\description{
Convert power to decibel and decibel to power.
}
\examples{
db <- pow2db(c(0, 10, 100))
pow <- db2pow(c(-10, 0, 10))

}
\author{
P. Sudeepam\cr
Conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
