% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinewave.R
\name{sinewave}
\alias{sinewave}
\title{Sine wave}
\usage{
sinewave(m, n = m, d = 0)
}
\arguments{
\item{m}{desired length of the generated series, specified as a positive
integer.}

\item{n}{rate, of the generated series, specified as a positive integer.
Default: \code{m}.}

\item{d}{delay, specified as a positive integer. Default: 0.}
}
\value{
Sine wave, returned as a vector of length \code{m}.
}
\description{
Generate a discrete sine wave.
}
\examples{
plot(sinewave(100, 10), type = "l")

}
\author{
Friedrich Leisch.\cr
Conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@gmail.com}.
}
